Info
This is a clone of a clone of a Gameboy which I have modified to have a sync pulse so it can sync TD-3, RD-6, Pocket Operators, Volcas, etc... 

I think this would be a cool project for the upcoming hackable Arduboy Mini, which I am keen to get.

Sync:
I send a 15ms pulse at the start of each note to pin PD2 (digital pin 0) which is connected to the sync out socket. (Tip not ring to match the Pocket Operators spec). I am using one trigger/pulse per step on the TD-3 and RD-6

With this program, which I call “Sync Master” there are 16 pulses in a sequence. These can be spaced as you like to produce different rhythms. In this Jam I focus on swing and a weird swing?

The “B” button stops/starts the triggers/pulses. The “A” button chooses the function of the D buttons which can do the following. 

There are the following possibilities when using the program on an Arduboy:
- Modify each step length/gate individually
- Make a global adjustment to all steps
- Instant patterns based on swing, straight, Sine wave, Ramp and a random length
- Mutations - where each step randomly changes
- Sequence Length - Change the length of the pattern to less than 16

In action here:
https://www.youtube.com/watch?v=FBHi6AkfH_4